requirements.txt
scikitlearn
numpy
panda
matplotlib
seaborn